Phase 4.1: UDP transport and endpoint enumeration
The second piece of direct paths: a socket that can send a packet straight at
a peer, and a list of addresses worth telling a peer to try.

Two sockets rather than one. A single IPv6 socket with V6ONLY off carries IPv4
through v4-mapped addresses, and then every address in the program has two
spellings and the bugs live in the seam between them. Separate sockets cost a
descriptor and remove that class outright. Both bind the same port where the
operating system allows, so a peer told "reach me on N" finds us there
whichever family it picks.

tc_endpoint moved to its own header. Two modules need it now, and having the
UDP transport include the STUN header to get an address type was backwards.

THE CANDIDATE FILTER IS THE PART THAT REPAYS TESTING

tc_endpoint_is_candidate decides what a peer spends round trips probing, and
both kinds of mistake are quiet. A wrongly accepted address is a probe that
can never arrive; a wrongly rejected one is a direct path never found. Neither
shows up as a failure -- the tunnel just stays on the relay, working, slower,
for reasons nobody investigates.

So the table is explicit about why each rejection is a rejection, and the
boundaries are tested from both sides: 100.63 and 100.128 are candidates while
the whole of 100.64/10 is not; febf:: is link-local and fec0:: is not, which
is the case a byte comparison instead of a mask gets wrong.

Two decisions worth stating:

  RFC 1918 and unique-local addresses stay. Two peers on one LAN reaching each
  other directly is the best path there is, and it is exactly what a filter
  written with only the public internet in mind would throw away.

  100.64/10 goes. It is carrier-grade NAT, and also the range Tailscale uses
  for its own nodes. Either way a peer cannot reach us there: behind a carrier
  NAT nobody can, and on a tailnet the address belongs to a different system.

VERIFICATION

test_udp, 69 checks: the filter table above, and the socket half over loopback
-- a datagram each way, the sender's address and port arriving with it (which
is the thing a direct path is built on), silence reported as a timeout rather
than an error, and a requested port honoured by both families.

tests/livestun.c now opens the real UDP socket first and asks STUN through
*it*, rather than through a throwaway. That matters: a mapping discovered for
a port nothing listens on is an answer that looks right and is useless. It
also lists the local candidates, which no unit test can asserts anything about
-- it cannot know what addresses this machine has.

Live, on this machine: three local candidates, and both DERP regions report
the same public address and port.

Also fixes three -Wconversion warnings cosmocc found in code added earlier in
Phase 3 (sockaddr family assignments and open()'s flags). The host build had
not flagged them; both toolchains are clean again.

All 19 suites pass under cosmocc (fat) and host gcc with ASan+UBSan.

+++ src/cli/main.c
@@ -748,7 +748,7 @@ static int dial_localhost(uint16_t port)
 {
 	struct sockaddr_in v4;
 	memset(&v4, 0, sizeof v4);
-	v4.sin_family = AF_INET;
+	v4.sin_family = (uint16_t)AF_INET;
 	v4.sin_port = htons(port);
 	v4.sin_addr.s_addr = htonl(INADDR_LOOPBACK);
 
@@ -761,7 +761,7 @@ static int dial_localhost(uint16_t port)
 
 	struct sockaddr_in6 v6;
 	memset(&v6, 0, sizeof v6);
-	v6.sin6_family = AF_INET6;
+	v6.sin6_family = (uint16_t)AF_INET6;
 	v6.sin6_port = htons(port);
 	v6.sin6_addr = in6addr_loopback;
 
@@ -1415,7 +1415,7 @@ static int cmd_genkey(const char *key_spec, bool client, bool force,
 	/* 0600 from the moment it exists, rather than created and then chmod'd:
 	 * the whole file is secret, including the pre-shared key that lives
 	 * under the misleading name "Public". */
-	int fd = open(path, O_WRONLY | O_CREAT | O_TRUNC, 0600);
+	int fd = open(path, (int)(O_WRONLY | O_CREAT | O_TRUNC), 0600);
 	if (fd < 0) {
 		fprintf(stderr, "tailcat-c: cannot write %s: %s\n", path,
 		        strerror(errno));
@@ -1695,7 +1695,7 @@ static int bind_local(const char *bind_addr, uint16_t port, uint16_t *bound)
 {
 	struct sockaddr_in a;
 	memset(&a, 0, sizeof a);
-	a.sin_family = AF_INET;
+	a.sin_family = (uint16_t)AF_INET;
 	a.sin_port = htons(port);
 	if (inet_pton(AF_INET, bind_addr, &a.sin_addr) != 1) {
 		fprintf(stderr, "tailcat-c: \"%s\" is not an IPv4 address\n",
